MainsaCheck/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
supabase>=1.0.3
python-dotenv>=1.0.0
httpx>=0.24.0
email-validator>=2.0.0
reportlab>=4.0.0